Initial OSA zuul v3 role jobs
This patch implements an initial set of jobs intended to match
the current job execution method. It does not intend to improve
how the jobs are executed - only to replicate what is currently
in openstack-infra/openstack-zuul-jobs and provide the platform
to iterate on.

The meta-deps never appeared to be used, meaning that the
openstack_openrc role was never used either, and given that
vars are missing to make that work, the role dep has been
removed.

The requirement for httplib2 to use the get_uri module was
removed in Ansible 2.1 [1], so the installation of that for
functional testing has also been removed.

The requirements.txt was not used anywhere, so it's been
removed. The test-requirements.txt has been updated.

[1] http://docs.ansible.com/ansible/latest/uri_module.html#notes
Change-Id: I25010299a4b264ae34390e38915d6df953448ae5